lol_stats/requirements.txt
# python==3.10.3
django==4.0.3
django-filter==21.1	
django-graphql-jwt==0.3.4	
djongo==1.3.6	
dnspython==2.2.1	
graphene==2.1.9	
graphene-django==2.15.0	
graphene-mongo==0.2.13	
graphql-core==2.3.2	
graphql-relay==2.0.1	
mongoengine==0.24.1	
pymongo==3.12.0	
pytz==2022.1
django-cors-headers==3.11.0
